VideoParsing/requirements.txt
flask==3.1.0
gunicorn==23.0.0
google-cloud-storage==2.19.0
google-cloud-bigquery==3.27.0
google-genai==1.14.0
